ModelService_graduation-main/ModelService/Main/requirements.txt

# ========== 核心Web框架 ==========
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
starlette>=0.14.2

# ========== HTTP客户端 ==========
httpx>=0.20.0
requests>=2.26.0

# ========== 环境配置 ==========
python-dotenv>=0.19.0

# ========== 异步和网络 ==========
websockets>=10.0
aiofiles>=0.8.0

# ========== 数据处理 ==========
jieba>=0.42.1

# ========== 图像处理 ==========
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0

# ========== 机器学习和科学计算 ==========
torch>=1.9.0
torchvision>=0.10.0
scipy>=1.7.0
scikit-learn>=0.24.2

# ========== 数据格式处理 ==========
pandas>=1.3.0
matplotlib>=3.4.0

# ========== 数据库和缓存 ==========
faiss-cpu>=1.7.0
chromadb>=0.4.0

# ========== 文件处理 ==========
PyPDF2>=2.0.0
python-docx>=0.8.11
openpyxl>=3.0.0

# ========== 其他工具 ==========
colorama>=0.4.4
tqdm>=4.62.0

# ========== Django相关（可选，用于某些功能）==========
django>=3.2.0
channels>=3.0.0
djangorestframework>=3.12.0
django-cors-headers>=3.8.0

# ========== 可选：如果需要GPU加速 ==========
# torch>=1.9.0+cu111 --index-url https://download.pytorch.org/whl/cu111
# torchvision>=0.10.0+cu111 --index-url https://download.pytorch.org/whl/cu111 